Conflicts:
	Canvas.h

+++ Camera.h
@@ -1,11 +1,11 @@
 class Camera {
   public:
-    Camera (double locX, double locY, double locZ, double viewAngle, double viewDirX, double viewDirY, double viewDirZ) : 
+    Camera (float locX, float locY, float locZ, float viewAngle, float viewDirX, float viewDirY, float viewDirZ) : 
     viewAngle(viewAngle), 
     location(locX, locY, locZ), 
     viewDirection(viewDirX, viewDirY, viewDirZ)
     {}
     Coord location;
-    double viewAngle;
+    float viewAngle;
     Coord viewDirection;
 };
